move uninit_assumed_init to its own module
